adjust OfferModal folder structure
