+++ src/pages/OfferModal/OfferPage/MakeOffer/MakeOffer.jsx
