+++ src/pages/OfferModal/OfferPage/MakeOffer/MakeOffer.module.scss
