+++ src/pages/OfferModal/OfferPage/MakeOffer/index.js
