+++ src/pages/OfferModal/ProfilePage/ProfilePage.module.scss
@@ -23,4 +23,9 @@
   }
 }
 
-
+.back_button_wrapper{
+  position: absolute;
+  left: 20px;
+  display: flex;
+  align-items: center;
+}
