+++ src/pages/OfferModal/ProfilePage/SubPages/BankAccount/BankAccount.jsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react';
 
 import styles from './BankAccount.module.scss';
 
-import Button from '../../../../components/Button';
-import Input from '../../../../components/Input';
+import Button from '../../../../../components/Button';
+import Input from '../../../../../components/Input';
 import useForm from '../useForm';
 import FORM from './form';
-import onlyNumber from '../../../../utils/validators/input';
+import onlyNumber from '../../../../../utils/validators/input';
 
 export default function BankAccount({ storedValue, onSubmit }) {
   const form = useForm(FORM, storedValue);
