+++ src/pages/OfferModal/ProfilePage/SubPages/Photo/Photo.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../components/scss-partial/icon-url'; 
+@import '../../../../../components/scss-partial/icon-url'; 
 
 .form_wrapper{
   height: 300px;
